+++ src/app/components/todo/todo-list/todo-list.component.ts
@@ -1,8 +1,12 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, AfterViewInit, ViewChild, ElementRef } from '@angular/core';
 
 import { trigger, transition, style, animate } from '@angular/animations';
 import { TodoService } from 'src/app/services/todo.service';
 import { AuthService } from 'src/app/services/auth.service';
+import { Store } from '@ngrx/store';
+
+import { fromEvent } from 'rxjs';
+import { TodosState } from 'src/app/store';
 
 @Component({
   selector: 'todo-list',
@@ -29,11 +33,14 @@ import { AuthService } from 'src/app/services/auth.service';
   ]
 })
 export class TodoListComponent implements OnInit {
+
   todoTitle: string;
 
   constructor(
     public todoService : TodoService,
-    public authService: AuthService
+    public authService: AuthService,
+    private store: Store<TodosState>,
+
     ) { 
   }
 
@@ -46,6 +53,8 @@ export class TodoListComponent implements OnInit {
     this.todoService.addTodo(this.todoTitle);
     this.todoTitle = '';
   }
+
+