+++ src/main.ts
@@ -1,7 +1,10 @@
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { provideStore } from '@ngrx/store';
 
 import { AppModule } from './app/app.module';
+import { reducers } from './app/store';
 
 
-platformBrowserDynamic().bootstrapModule(AppModule)
+platformBrowserDynamic().bootstrapModule(
+  AppModule,)
   .catch(err => console.error(err));