+++ CHANGELOG.md
@@ -1,2 +1,14 @@
-#### 0.0.0 (2023-01-28)
+#### 0.0.0 (2023-02-07)
 This project is based on the [Angular Todo App](https://www.youtube.com/watch?v=nbFmHD4T04k&list=PLEhEHUEU3x5rMkFPBSvEhAok1WadjeWkG&index=1&t=2s) tutorial from Andre Madarang.
+
+##### New Features
+
+*  translate todos (67557882)
+*  add users info module (d49c1bba)
+*  todos, user auth, admin auth (a3e1f017)
+*  github actions (f8473517)
+
+##### Bug Fixes
+
+*  modify gitignore (c996ce70)
+